pyproject.toml
[tool.poetry]
name = "cf-token-miner"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.5.2"
pyautogui = "^0.9.54"
pytesseract = "^0.3.13"
redis = "^5.1.0"
pillow = "^10.4.0"
pyvirtualdisplay = "^3.0"
python-xlib = "^0.33"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.0"
alembic = "^1.13.3"
psycopg2 = "^2.9.9"
uvicorn = "^0.31.0"
email-validator = "^2.2.0"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
